+++ src/movie_file_fixer/poster_finder.py
@@ -12,10 +12,15 @@ import requests
 
 class PosterFinder:
     def __init__(
-        self, directory=None, metadata_filename="metadata.json", verbose=False
+        self,
+        directory=None,
+        metadata_filename="metadata.json",
+        dry_run=False,
+        verbose=False,
     ):
         self._directory = directory
         self._metadata_filename = metadata_filename
+        self._dry_run = dry_run
         self._verbose = verbose
         self._action_counter = 0
 
@@ -29,6 +34,8 @@ class PosterFinder:
         :param dict headers: A dictionary containing custom headers. Default only contains the `User-Agent`.
         :return requests.Response: A `requests.Response` object containing the file being requested.
         """
+        response = None
+
         if headers is None:
             headers = {
                 "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/74.0.3729.169 Safari/537.36"
@@ -38,7 +45,10 @@ class PosterFinder:
             print(f'[{self._action_counter}] [DOWNLOADING] [FILE] from [URL] "{url}"\n')
             self._action_counter += 1
 
-        return requests.get(url=url, headers=headers)
+        if not self._dry_run:
+            response = requests.get(url=url, headers=headers)
+
+        return response
 
     def get_posters(self, directory=None, metadata_filename=None):
         """
@@ -93,13 +103,17 @@ class PosterFinder:
 
                         response = self._download(url=poster_url)
 
-                        if response.status_code == 200:
-                            if self._verbose:
-                                print("[DOWNLOAD COMPLETE]\n")
-                                print(
-                                    f'[{self._action_counter}] [WRITING FILE] -> "{poster_filepath}"\n'
-                                )
-                                self._action_counter += 1
+                        if not self._dry_run:
+                            if response.status_code == 200:
+                                if self._verbose:
+                                    print("[DOWNLOAD COMPLETE]\n")
+                                    print(
+                                        f'[{self._action_counter}] [WRITING FILE] -> "{poster_filepath}"\n'
+                                    )
+                                    self._action_counter += 1
 
-                            with open(poster_filepath, "wb") as outfile:
-                                outfile.write(response.content)
+                                with open(poster_filepath, "wb") as outfile:
+                                    outfile.write(response.content)
+                        else:
+                            print("[DRY MODE ACTIVATED, POSTER NOT DOWNLOADED]\n")
+                            self._action_counter += 1
